Support Facter 1.7.x in spec tests
The facter project has renamed `loader` into `internal_loader` which
breaks the ldom spec test. The ldom fact needs to be explicitly loaded
by facter because the autoloader cannot find it (the facts are named
different than the filename they are stored in).

Use internal_loader when available otherwise suppose we are on facter <
1.6.x and use loader.

For reference: The change that caused the spec failures was introduced
in commit 082dba6 of the facter project.